% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurri.R
\name{read_ext_tracks}
\alias{read_ext_tracks}
\title{read in data}
\usage{
read_ext_tracks(
  file = "ebtrk_atlc_1988_2015.txt",
  ...,
  widths = NULL,
  colnames = NULL,
  degW = TRUE
)
}
\description{
This function is to import the data set.
}
